refactor product, cart and success pages
